+++ deno/rest/v10/channel.ts
@@ -26,6 +26,7 @@ import type {
 	ChannelFlags,
 } from '../../payloads/v10/mod.ts';
 import type { AddUndefinedToPossiblyUndefinedPropertiesOfInterface, StrictPartial } from '../../utils/internals.ts';
+import type { RESTAPIPollCreate } from './poll.ts';
 
 export interface APIChannelPatchOverwrite extends RESTPutAPIChannelPermissionJSONBody {
 	id: Snowflake;
@@ -236,8 +237,10 @@ export type RESTGetAPIChannelMessageResult = APIMessage;
 /**
  * https://discord.com/developers/docs/resources/channel#message-reference-object-message-reference-structure
  */
-export type APIMessageReferenceSend = StrictPartial<APIMessageReference> &
-	AddUndefinedToPossiblyUndefinedPropertiesOfInterface<Required<Pick<APIMessageReference, 'message_id'>>> & {
+export type APIMessageReferenceSend = AddUndefinedToPossiblyUndefinedPropertiesOfInterface<
+	Required<Pick<APIMessageReference, 'message_id'>>
+> &
+	StrictPartial<APIMessageReference> & {
 		/**
 		 * Whether to error if the referenced message doesn't exist instead of sending as a normal (non-reply) message
 		 *
@@ -309,7 +312,7 @@ export interface RESTPostAPIChannelMessageJSONBody {
 	 *
 	 * See https://discord.com/developers/docs/resources/sticker#sticker-object
 	 */
-	sticker_ids?: [Snowflake] | [Snowflake, Snowflake] | [Snowflake, Snowflake, Snowflake] | undefined;
+	sticker_ids?: [Snowflake, Snowflake, Snowflake] | [Snowflake, Snowflake] | [Snowflake] | undefined;
 	/**
 	 * Attachment objects with filename and description
 	 */
@@ -318,19 +321,28 @@ export interface RESTPostAPIChannelMessageJSONBody {
 	 * Message flags combined as a bitfield
 	 */
 	flags?: MessageFlags | undefined;
+	/**
+	 * If `true` and nonce is present, it will be checked for uniqueness in the past few minutes.
+	 * If another message was created by the same author with the same nonce, that message will be returned and no new message will be created.
+	 */
+	enforce_nonce?: boolean | undefined;
+	/**
+	 * A poll!
+	 */
+	poll?: RESTAPIPollCreate | undefined;
 }
 
 /**
  * https://discord.com/developers/docs/resources/channel#create-message
  */
 export type RESTPostAPIChannelMessageFormDataBody =
-	| ({
+	| (Record<`files[${bigint}]`, unknown> & {
 			/**
 			 * JSON stringified message body
 			 */
 			payload_json?: string | undefined;
-	  } & Record<`files[${bigint}]`, unknown>)
-	| (RESTPostAPIChannelMessageJSONBody & Record<`files[${bigint}]`, unknown>);
+	  })
+	| (Record<`files[${bigint}]`, unknown> & RESTPostAPIChannelMessageJSONBody);
 
 /**
  * https://discord.com/developers/docs/resources/channel#create-message
@@ -437,13 +449,13 @@ export interface RESTPatchAPIChannelMessageJSONBody {
  * https://discord.com/developers/docs/resources/channel#edit-message
  */
 export type RESTPatchAPIChannelMessageFormDataBody =
-	| ({
+	| (Record<`files[${bigint}]`, unknown> & {
 			/**
 			 * JSON stringified message body
 			 */
 			payload_json?: string | undefined;
-	  } & Record<`files[${bigint}]`, unknown>)
-	| (RESTPatchAPIChannelMessageJSONBody & Record<`files[${bigint}]`, unknown>);
+	  })
+	| (Record<`files[${bigint}]`, unknown> & RESTPatchAPIChannelMessageJSONBody);
 
 /**
  * https://discord.com/developers/docs/resources/channel#edit-message