+++ deno/rest/v9/mod.ts
@@ -12,6 +12,7 @@ export * from './guildScheduledEvent.ts';
 export * from './interactions.ts';
 export * from './invite.ts';
 export * from './oauth2.ts';
+export * from './poll.ts';
 export * from './stageInstance.ts';
 export * from './sticker.ts';
 export * from './template.ts';
@@ -463,6 +464,22 @@ export const Routes = {
 
 	/**
 	 * Route for:
+	 * - GET `/channels/{channel.id}/polls/{message.id}/answers/{answer_id}`
+	 */
+	pollAnswerVoters(channelId: Snowflake, messageId: Snowflake, answerId: number) {
+		return `/channels/${channelId}/polls/${messageId}/answers/${answerId}` as const;
+	},
+
+	/**
+	 * Route for:
+	 * - POST `/channels/{channel.id}/polls/{message.id}/expire`
+	 */
+	expirePoll(channelId: Snowflake, messageId: Snowflake) {
+		return `/channels/${channelId}/polls/${messageId}/expire` as const;
+	},
+
+	/**
+	 * Route for:
 	 * - POST `/channels/{channel.id}/threads`
 	 * - POST `/channels/{channel.id}/messages/{message.id}/threads`
 	 */
@@ -473,7 +490,9 @@ export const Routes = {
 
 		parts.push('threads');
 
-		return parts.join('/') as `/channels/${Snowflake}/threads` | `/channels/${Snowflake}/messages/${Snowflake}/threads`;
+		return parts.join('/') as
+			| `/channels/${Snowflake}/messages/${Snowflake}/threads`
+			| `/channels/${Snowflake}/threads`;
 	},
 
 	/**
@@ -491,7 +510,7 @@ export const Routes = {
 	 * - GET `/channels/{channel.id}/threads/archived/public`
 	 * - GET `/channels/{channel.id}/threads/archived/private`
 	 */
-	channelThreads(channelId: Snowflake, archived?: 'public' | 'private') {
+	channelThreads(channelId: Snowflake, archived?: 'private' | 'public') {
 		const parts = ['', 'channels', channelId, 'threads'];
 
 		if (archived) parts.push('archived', archived);
@@ -499,7 +518,7 @@ export const Routes = {
 
 		return parts.join('/') as
 			| `/channels/${Snowflake}/threads/active`
-			| `/channels/${Snowflake}/threads/archived/${'public' | 'private'}`;
+			| `/channels/${Snowflake}/threads/archived/${'private' | 'public'}`;
 	},
 
 	/**
@@ -525,8 +544,8 @@ export const Routes = {
 		if (userId) parts.push(userId);
 
 		return parts.join('/') as
-			| `/channels/${Snowflake}/thread-members`
-			| `/channels/${Snowflake}/thread-members/${Snowflake | '@me'}`;
+			| `/channels/${Snowflake}/thread-members/${Snowflake | '@me'}`
+			| `/channels/${Snowflake}/thread-members`;
 	},
 
 	/**
@@ -632,7 +651,7 @@ export const Routes = {
 
 		if (webhookToken) parts.push(webhookToken);
 
-		return parts.join('/') as `/webhooks/${Snowflake}` | `/webhooks/${Snowflake}/${string}`;
+		return parts.join('/') as `/webhooks/${Snowflake}/${string}` | `/webhooks/${Snowflake}`;
 	},
 
 	/**
@@ -917,6 +936,7 @@ export const Routes = {
 	currentApplication() {
 		return '/applications/@me' as const;
 	},
+
 	/**
 	 * Route for:
 	 * - GET `/applications/{application.id}/entitlements`
@@ -925,6 +945,7 @@ export const Routes = {
 	entitlements(applicationId: Snowflake) {
 		return `/applications/${applicationId}/entitlements` as const;
 	},
+
 	/**
 	 * Route for:
 	 * - DELETE `/applications/{application.id}/entitlements/{entitlement.id}`
@@ -932,6 +953,7 @@ export const Routes = {
 	entitlement(applicationId: Snowflake, entitlementId: Snowflake) {
 		return `/applications/${applicationId}/entitlements/${entitlementId}` as const;
 	},
+
 	/**
 	 * Route for:
 	 * - GET `/applications/{application.id}/skus`
@@ -939,9 +961,25 @@ export const Routes = {
 	skus(applicationId: Snowflake) {
 		return `/applications/${applicationId}/skus` as const;
 	},
+
+	/**
+	 * Route for:
+	 * - POST `/guilds/{guild.id}/bulk-ban`
+	 */
+	guildBulkBan(guildId: Snowflake) {
+		return `/guilds/${guildId}/bulk-ban` as const;
+	},
 };
 
 export const StickerPackApplicationId = '710982414301790216';
+
+export enum ImageFormat {
+	JPEG = 'jpeg',
+	PNG = 'png',
+	WebP = 'webp',
+	GIF = 'gif',
+	Lottie = 'json',
+}
 
 export const CDNRoutes = {
 	/**
@@ -952,7 +990,7 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP, GIF
 	 */
-	emoji(emojiId: Snowflake, format: EmojiFormat) {
+	emoji<Format extends EmojiFormat>(emojiId: Snowflake, format: Format) {
 		return `/emojis/${emojiId}.${format}` as const;
 	},
 
@@ -964,7 +1002,7 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP, GIF
 	 */
-	guildIcon(guildId: Snowflake, guildIcon: string, format: GuildIconFormat) {
+	guildIcon<Format extends GuildIconFormat>(guildId: Snowflake, guildIcon: string, format: Format) {
 		return `icons/${guildId}/${guildIcon}.${format}` as const;
 	},
 
@@ -974,7 +1012,7 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP
 	 */
-	guildSplash(guildId: Snowflake, guildSplash: string, format: GuildSplashFormat) {
+	guildSplash<Format extends GuildSplashFormat>(guildId: Snowflake, guildSplash: string, format: Format) {
 		return `/splashes/${guildId}/${guildSplash}.${format}` as const;
 	},
 
@@ -984,7 +1022,11 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP
 	 */
-	guildDiscoverySplash(guildId: Snowflake, guildDiscoverySplash: string, format: GuildDiscoverySplashFormat) {
+	guildDiscoverySplash<Format extends GuildDiscoverySplashFormat>(
+		guildId: Snowflake,
+		guildDiscoverySplash: string,
+		format: Format,
+	) {
 		return `/discovery-splashes/${guildId}/${guildDiscoverySplash}.${format}` as const;
 	},
 
@@ -996,7 +1038,7 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP, GIF
 	 */
-	guildBanner(guildId: Snowflake, guildBanner: string, format: GuildBannerFormat) {
+	guildBanner<Format extends GuildBannerFormat>(guildId: Snowflake, guildBanner: string, format: Format) {
 		return `/banners/${guildId}/${guildBanner}.${format}` as const;
 	},
 
@@ -1008,7 +1050,7 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP, GIF
 	 */
-	userBanner(userId: Snowflake, userBanner: string, format: UserBannerFormat) {
+	userBanner<Format extends UserBannerFormat>(userId: Snowflake, userBanner: string, format: Format) {
 		return `/banners/${userId}/${userBanner}.${format}` as const;
 	},
 
@@ -1022,7 +1064,7 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extension: PNG
 	 */
-	defaultUserAvatar(index: DefaultUserAvatarAssets) {
+	defaultUserAvatar<Index extends DefaultUserAvatarAssets>(index: Index) {
 		return `/embed/avatars/${index}.png` as const;
 	},
 
@@ -1034,7 +1076,7 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP, GIF
 	 */
-	userAvatar(userId: Snowflake, userAvatar: string, format: UserAvatarFormat) {
+	userAvatar<Format extends UserAvatarFormat>(userId: Snowflake, userAvatar: string, format: Format) {
 		return `/avatars/${userId}/${userAvatar}.${format}` as const;
 	},
 
@@ -1046,7 +1088,12 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP, GIF
 	 */
-	guildMemberAvatar(guildId: Snowflake, userId: Snowflake, memberAvatar: string, format: GuildMemberAvatarFormat) {
+	guildMemberAvatar<Format extends GuildMemberAvatarFormat>(
+		guildId: Snowflake,
+		userId: Snowflake,
+		memberAvatar: string,
+		format: Format,
+	) {
 		return `/guilds/${guildId}/users/${userId}/avatars/${memberAvatar}.${format}` as const;
 	},
 
@@ -1066,7 +1113,11 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP
 	 */
-	applicationIcon(applicationId: Snowflake, applicationIcon: string, format: ApplicationIconFormat) {
+	applicationIcon<Format extends ApplicationIconFormat>(
+		applicationId: Snowflake,
+		applicationIcon: string,
+		format: Format,
+	) {
 		return `/app-icons/${applicationId}/${applicationIcon}.${format}` as const;
 	},
 
@@ -1076,7 +1127,11 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP
 	 */
-	applicationCover(applicationId: Snowflake, applicationCoverImage: string, format: ApplicationCoverFormat) {
+	applicationCover<Format extends ApplicationCoverFormat>(
+		applicationId: Snowflake,
+		applicationCoverImage: string,
+		format: Format,
+	) {
 		return `/app-icons/${applicationId}/${applicationCoverImage}.${format}` as const;
 	},
 
@@ -1086,7 +1141,11 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP
 	 */
-	applicationAsset(applicationId: Snowflake, applicationAssetId: string, format: ApplicationAssetFormat) {
+	applicationAsset<Format extends ApplicationAssetFormat>(
+		applicationId: Snowflake,
+		applicationAssetId: string,
+		format: Format,
+	) {
 		return `/app-assets/${applicationId}/${applicationAssetId}.${format}` as const;
 	},
 
@@ -1096,11 +1155,11 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP
 	 */
-	achievementIcon(
+	achievementIcon<Format extends AchievementIconFormat>(
 		applicationId: Snowflake,
 		achievementId: Snowflake,
 		achievementIconHash: string,
-		format: AchievementIconFormat,
+		format: Format,
 	) {
 		return `/app-assets/${applicationId}/achievements/${achievementId}/icons/${achievementIconHash}.${format}` as const;
 	},
@@ -1111,7 +1170,7 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP
 	 */
-	stickerPackBanner(stickerPackBannerAssetId: Snowflake, format: StickerPackBannerFormat) {
+	stickerPackBanner<Format extends StickerPackBannerFormat>(stickerPackBannerAssetId: Snowflake, format: Format) {
 		return `/app-assets/${StickerPackApplicationId}/store/${stickerPackBannerAssetId}.${format}` as const;
 	},
 
@@ -1121,8 +1180,12 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP
 	 */
-	storePageAsset(applicationId: Snowflake, assetId: string) {
-		return `/app-assets/${applicationId}/store/${assetId}.png` as const;
+	storePageAsset<Format extends StorePageAssetFormat = ImageFormat.PNG>(
+		applicationId: Snowflake,
+		assetId: string,
+		format: Format = ImageFormat.PNG as Format,
+	) {
+		return `/app-assets/${applicationId}/store/${assetId}.${format}` as const;
 	},
 
 	/**
@@ -1131,7 +1194,7 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP
 	 */
-	teamIcon(teamId: Snowflake, teamIcon: string, format: TeamIconFormat) {
+	teamIcon<Format extends TeamIconFormat>(teamId: Snowflake, teamIcon: string, format: Format) {
 		return `/team-icons/${teamId}/${teamIcon}.${format}` as const;
 	},
 
@@ -1141,7 +1204,7 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, Lottie, GIF
 	 */
-	sticker(stickerId: Snowflake, format: StickerFormat) {
+	sticker<Format extends StickerFormat>(stickerId: Snowflake, format: Format) {
 		return `/stickers/${stickerId}.${format}` as const;
 	},
 
@@ -1151,7 +1214,7 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP
 	 */
-	roleIcon(roleId: Snowflake, roleIcon: string, format: RoleIconFormat) {
+	roleIcon<Format extends RoleIconFormat>(roleId: Snowflake, roleIcon: string, format: Format) {
 		return `/role-icons/${roleId}/${roleIcon}.${format}` as const;
 	},
 
@@ -1161,10 +1224,10 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP
 	 */
-	guildScheduledEventCover(
+	guildScheduledEventCover<Format extends GuildScheduledEventCoverFormat>(
 		guildScheduledEventId: Snowflake,
 		guildScheduledEventCoverImage: string,
-		format: GuildScheduledEventCoverFormat,
+		format: Format,
 	) {
 		return `/guild-events/${guildScheduledEventId}/${guildScheduledEventCoverImage}.${format}` as const;
 	},
@@ -1175,7 +1238,12 @@ export const CDNRoutes = {
 	 *
 	 * This route supports the extensions: PNG, JPEG, WebP, GIF
 	 */
-	guildMemberBanner(guildId: Snowflake, userId: Snowflake, guildMemberBanner: string, format: GuildMemberBannerFormat) {
+	guildMemberBanner<Format extends GuildMemberBannerFormat>(
+		guildId: Snowflake,
+		userId: Snowflake,
+		guildMemberBanner: string,
+		format: Format,
+	) {
 		return `/guilds/${guildId}/users/${userId}/banners/${guildMemberBanner}.${format}` as const;
 	},
 };
@@ -1184,31 +1252,24 @@ export type DefaultUserAvatarAssets = 0 | 1 | 2 | 3 | 4 | 5;
 
 export type EmojiFormat = Exclude<ImageFormat, ImageFormat.Lottie>;
 export type GuildIconFormat = Exclude<ImageFormat, ImageFormat.Lottie>;
-export type GuildSplashFormat = Exclude<ImageFormat, ImageFormat.Lottie | ImageFormat.GIF>;
-export type GuildDiscoverySplashFormat = Exclude<ImageFormat, ImageFormat.Lottie | ImageFormat.GIF>;
+export type GuildSplashFormat = Exclude<ImageFormat, ImageFormat.GIF | ImageFormat.Lottie>;
+export type GuildDiscoverySplashFormat = Exclude<ImageFormat, ImageFormat.GIF | ImageFormat.Lottie>;
 export type GuildBannerFormat = Exclude<ImageFormat, ImageFormat.Lottie>;
 export type UserBannerFormat = Exclude<ImageFormat, ImageFormat.Lottie>;
 export type DefaultUserAvatar = Extract<ImageFormat, ImageFormat.PNG>;
 export type UserAvatarFormat = Exclude<ImageFormat, ImageFormat.Lottie>;
 export type GuildMemberAvatarFormat = Exclude<ImageFormat, ImageFormat.Lottie>;
-export type ApplicationIconFormat = Exclude<ImageFormat, ImageFormat.Lottie | ImageFormat.GIF>;
-export type ApplicationCoverFormat = Exclude<ImageFormat, ImageFormat.Lottie | ImageFormat.GIF>;
-export type ApplicationAssetFormat = Exclude<ImageFormat, ImageFormat.Lottie | ImageFormat.GIF>;
-export type AchievementIconFormat = Exclude<ImageFormat, ImageFormat.Lottie | ImageFormat.GIF>;
-export type StickerPackBannerFormat = Exclude<ImageFormat, ImageFormat.Lottie | ImageFormat.GIF>;
-export type TeamIconFormat = Exclude<ImageFormat, ImageFormat.Lottie | ImageFormat.GIF>;
-export type StickerFormat = Extract<ImageFormat, ImageFormat.PNG | ImageFormat.Lottie | ImageFormat.GIF>;
-export type RoleIconFormat = Exclude<ImageFormat, ImageFormat.Lottie | ImageFormat.GIF>;
-export type GuildScheduledEventCoverFormat = Exclude<ImageFormat, ImageFormat.Lottie | ImageFormat.GIF>;
+export type ApplicationIconFormat = Exclude<ImageFormat, ImageFormat.GIF | ImageFormat.Lottie>;
+export type ApplicationCoverFormat = Exclude<ImageFormat, ImageFormat.GIF | ImageFormat.Lottie>;
+export type ApplicationAssetFormat = Exclude<ImageFormat, ImageFormat.GIF | ImageFormat.Lottie>;
+export type AchievementIconFormat = Exclude<ImageFormat, ImageFormat.GIF | ImageFormat.Lottie>;
+export type StickerPackBannerFormat = Exclude<ImageFormat, ImageFormat.GIF | ImageFormat.Lottie>;
+export type StorePageAssetFormat = Exclude<ImageFormat, ImageFormat.GIF | ImageFormat.Lottie>;
+export type TeamIconFormat = Exclude<ImageFormat, ImageFormat.GIF | ImageFormat.Lottie>;
+export type StickerFormat = Extract<ImageFormat, ImageFormat.GIF | ImageFormat.Lottie | ImageFormat.PNG>;
+export type RoleIconFormat = Exclude<ImageFormat, ImageFormat.GIF | ImageFormat.Lottie>;
+export type GuildScheduledEventCoverFormat = Exclude<ImageFormat, ImageFormat.GIF | ImageFormat.Lottie>;
 export type GuildMemberBannerFormat = Exclude<ImageFormat, ImageFormat.Lottie>;
-
-export enum ImageFormat {
-	JPEG = 'jpeg',
-	PNG = 'png',
-	WebP = 'webp',
-	GIF = 'gif',
-	Lottie = 'json',
-}
 
 export interface CDNQuery {
 	/**