+++ payloads/v10/auditLog.ts
@@ -17,6 +17,7 @@ import type {
 	GuildDefaultMessageNotifications,
 	GuildExplicitContentFilter,
 	GuildMFALevel,
+	GuildSystemChannelFlags,
 	GuildVerificationLevel,
 	IntegrationExpireBehavior,
 } from './guild';
@@ -335,82 +336,83 @@ export type AuditLogRuleTriggerType = `${AutoModerationRuleTriggerType}`;
  * https://discord.com/developers/docs/resources/audit-log#audit-log-change-object-audit-log-change-structure
  */
 export type APIAuditLogChange =
-	| APIAuditLogChangeKeyName
-	| APIAuditLogChangeKeyDescription
-	| APIAuditLogChangeKeyIconHash
-	| APIAuditLogChangeKeyImageHash
-	| APIAuditLogChangeKeySplashHash
-	| APIAuditLogChangeKeyDiscoverySplashHash
-	| APIAuditLogChangeKeyBannerHash
-	| APIAuditLogChangeKeyOwnerId
-	| APIAuditLogChangeKeyRegion
-	| APIAuditLogChangeKeyPreferredLocale
-	| APIAuditLogChangeKeyAFKChannelId
-	| APIAuditLogChangeKeyAFKTimeout
-	| APIAuditLogChangeKeyRulesChannelId
-	| APIAuditLogChangeKeyPublicUpdatesChannelId
-	| APIAuditLogChangeKeyMFALevel
-	| APIAuditLogChangeKeyVerificationLevel
-	| APIAuditLogChangeKeyExplicitContentFilter
-	| APIAuditLogChangeKeyDefaultMessageNotifications
-	| APIAuditLogChangeKeyVanityURLCode
 	| APIAuditLogChangeKey$Add
 	| APIAuditLogChangeKey$Remove
-	| APIAuditLogChangeKeyPruneDeleteDays
-	| APIAuditLogChangeKeyWidgetEnabled
-	| APIAuditLogChangeKeyWidgetChannelId
-	| APIAuditLogChangeKeySystemChannelId
-	| APIAuditLogChangeKeyPosition
-	| APIAuditLogChangeKeyTopic
-	| APIAuditLogChangeKeyBitrate
-	| APIAuditLogChangeKeyPermissionOverwrites
-	| APIAuditLogChangeKeyNSFW
-	| APIAuditLogChangeKeyApplicationId
-	| APIAuditLogChangeKeyRateLimitPerUser
-	| APIAuditLogChangeKeyPermissions
-	| APIAuditLogChangeKeyColor
-	| APIAuditLogChangeKeyHoist
-	| APIAuditLogChangeKeyMentionable
+	| APIAuditLogChangeKeyActions
+	| APIAuditLogChangeKeyAFKChannelId
+	| APIAuditLogChangeKeyAFKTimeout
 	| APIAuditLogChangeKeyAllow
-	| APIAuditLogChangeKeyDeny
-	| APIAuditLogChangeKeyCode
-	| APIAuditLogChangeKeyChannelId
-	| APIAuditLogChangeKeyInviterId
-	| APIAuditLogChangeKeyMaxUses
-	| APIAuditLogChangeKeyUses
-	| APIAuditLogChangeKeyMaxAge
-	| APIAuditLogChangeKeyTemporary
-	| APIAuditLogChangeKeyDeaf
-	| APIAuditLogChangeKeyMute
-	| APIAuditLogChangeKeyNick
+	| APIAuditLogChangeKeyApplicationId
+	| APIAuditLogChangeKeyArchived
+	| APIAuditLogChangeKeyAsset
+	| APIAuditLogChangeKeyAutoArchiveDuration
+	| APIAuditLogChangeKeyAvailable
 	| APIAuditLogChangeKeyAvatarHash
-	| APIAuditLogChangeKeyId
-	| APIAuditLogChangeKeyType
+	| APIAuditLogChangeKeyBannerHash
+	| APIAuditLogChangeKeyBitrate
+	| APIAuditLogChangeKeyChannelId
+	| APIAuditLogChangeKeyCode
+	| APIAuditLogChangeKeyColor
+	| APIAuditLogChangeKeyCommunicationDisabledUntil
+	| APIAuditLogChangeKeyDeaf
+	| APIAuditLogChangeKeyDefaultAutoArchiveDuration
+	| APIAuditLogChangeKeyDefaultMessageNotifications
+	| APIAuditLogChangeKeyDeny
+	| APIAuditLogChangeKeyDescription
+	| APIAuditLogChangeKeyDiscoverySplashHash
+	| APIAuditLogChangeKeyEnabled
 	| APIAuditLogChangeKeyEnableEmoticons
+	| APIAuditLogChangeKeyEntityType
+	| APIAuditLogChangeKeyEventType
+	| APIAuditLogChangeKeyExemptChannels
+	| APIAuditLogChangeKeyExemptRoles
 	| APIAuditLogChangeKeyExpireBehavior
 	| APIAuditLogChangeKeyExpireGracePeriod
-	| APIAuditLogChangeKeyUserLimit
-	| APIAuditLogChangeKeyPrivacyLevel
-	| APIAuditLogChangeKeyTags
+	| APIAuditLogChangeKeyExplicitContentFilter
 	| APIAuditLogChangeKeyFormatType
-	| APIAuditLogChangeKeyAsset
-	| APIAuditLogChangeKeyAvailable
 	| APIAuditLogChangeKeyGuildId
-	| APIAuditLogChangeKeyArchived
-	| APIAuditLogChangeKeyLocked
-	| APIAuditLogChangeKeyAutoArchiveDuration
-	| APIAuditLogChangeKeyDefaultAutoArchiveDuration
-	| APIAuditLogChangeKeyEntityType
-	| APIAuditLogChangeKeyStatus
+	| APIAuditLogChangeKeyHoist
+	| APIAuditLogChangeKeyIconHash
+	| APIAuditLogChangeKeyId
+	| APIAuditLogChangeKeyImageHash
+	| APIAuditLogChangeKeyInviterId
 	| APIAuditLogChangeKeyLocation
-	| APIAuditLogChangeKeyCommunicationDisabledUntil
-	| APIAuditLogChangeKeyTriggerType
-	| APIAuditLogChangeKeyEventType
+	| APIAuditLogChangeKeyLocked
+	| APIAuditLogChangeKeyMaxAge
+	| APIAuditLogChangeKeyMaxUses
+	| APIAuditLogChangeKeyMentionable
+	| APIAuditLogChangeKeyMFALevel
+	| APIAuditLogChangeKeyMute
+	| APIAuditLogChangeKeyName
+	| APIAuditLogChangeKeyNick
+	| APIAuditLogChangeKeyNSFW
+	| APIAuditLogChangeKeyOwnerId
+	| APIAuditLogChangeKeyPermissionOverwrites
+	| APIAuditLogChangeKeyPermissions
+	| APIAuditLogChangeKeyPosition
+	| APIAuditLogChangeKeyPreferredLocale
+	| APIAuditLogChangeKeyPrivacyLevel
+	| APIAuditLogChangeKeyPruneDeleteDays
+	| APIAuditLogChangeKeyPublicUpdatesChannelId
+	| APIAuditLogChangeKeyRateLimitPerUser
+	| APIAuditLogChangeKeyRegion
+	| APIAuditLogChangeKeyRulesChannelId
+	| APIAuditLogChangeKeySplashHash
+	| APIAuditLogChangeKeyStatus
+	| APIAuditLogChangeKeySystemChannelFlags
+	| APIAuditLogChangeKeySystemChannelId
+	| APIAuditLogChangeKeyTags
+	| APIAuditLogChangeKeyTemporary
+	| APIAuditLogChangeKeyTopic
 	| APIAuditLogChangeKeyTriggerMetadata
-	| APIAuditLogChangeKeyActions
-	| APIAuditLogChangeKeyEnabled
-	| APIAuditLogChangeKeyExemptRoles
-	| APIAuditLogChangeKeyExemptChannels;
+	| APIAuditLogChangeKeyTriggerType
+	| APIAuditLogChangeKeyType
+	| APIAuditLogChangeKeyUserLimit
+	| APIAuditLogChangeKeyUses
+	| APIAuditLogChangeKeyVanityURLCode
+	| APIAuditLogChangeKeyVerificationLevel
+	| APIAuditLogChangeKeyWidgetChannelId
+	| APIAuditLogChangeKeyWidgetEnabled;
 
 /**
  * Returned when an entity's name is changed
@@ -542,6 +544,14 @@ export type APIAuditLogChangeKeyWidgetEnabled = AuditLogChangeData<'widget_enabl
  * Returned when a guild's widget_channel_id is changed
  */
 export type APIAuditLogChangeKeyWidgetChannelId = AuditLogChangeData<'widget_channel_id', Snowflake>;
+
+/**
+ * Returned when a guild's system_channel_flags is changed
+ */
+export type APIAuditLogChangeKeySystemChannelFlags = AuditLogChangeData<
+	'system_channel_flags',
+	GuildSystemChannelFlags
+>;
 
 /**
  * Returned when a guild's system_channel_id is changed