+++ payloads/v10/guildScheduledEvent.ts
@@ -90,9 +90,9 @@ export interface APIExternalGuildScheduledEvent
  * https://discord.com/developers/docs/resources/guild-scheduled-event#guild-scheduled-event-object-guild-scheduled-event-structure
  */
 export type APIGuildScheduledEvent =
+	| APIExternalGuildScheduledEvent
 	| APIStageInstanceGuildScheduledEvent
-	| APIVoiceGuildScheduledEvent
-	| APIExternalGuildScheduledEvent;
+	| APIVoiceGuildScheduledEvent;
 
 /**
  * https://discord.com/developers/docs/resources/guild-scheduled-event#guild-scheduled-event-object-guild-scheduled-event-entity-metadata