+++ payloads/v8/interactions.ts
@@ -34,11 +34,11 @@ export * from './_interactions/responses';
  * @deprecated API and gateway v8 are deprecated and the types will not receive further updates, please update to v10.
  */
 export type APIInteraction =
-	| APIPingInteraction
+	| APIApplicationCommandAutocompleteInteraction
 	| APIApplicationCommandInteraction
 	| APIMessageComponentInteraction
-	| APIApplicationCommandAutocompleteInteraction
-	| APIModalSubmitInteraction;
+	| APIModalSubmitInteraction
+	| APIPingInteraction;
 
 /**
  * https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object
@@ -46,9 +46,9 @@ export type APIInteraction =
  * @deprecated API and gateway v8 are deprecated and the types will not receive further updates, please update to v10.
  */
 export type APIDMInteraction =
+	| APIApplicationCommandAutocompleteDMInteraction
 	| APIApplicationCommandDMInteraction
 	| APIMessageComponentDMInteraction
-	| APIApplicationCommandAutocompleteDMInteraction
 	| APIModalSubmitDMInteraction;
 
 /**
@@ -57,7 +57,7 @@ export type APIDMInteraction =
  * @deprecated API and gateway v8 are deprecated and the types will not receive further updates, please update to v10.
  */
 export type APIGuildInteraction =
+	| APIApplicationCommandAutocompleteGuildInteraction
 	| APIApplicationCommandGuildInteraction
 	| APIMessageComponentGuildInteraction
-	| APIApplicationCommandAutocompleteGuildInteraction
 	| APIModalSubmitGuildInteraction;