+++ rest/common.ts
@@ -254,6 +254,7 @@ export enum RESTJSONErrorCodes {
 	NoUsersWithDiscordTagExist = 80_004,
 
 	ReactionWasBlocked = 90_001,
+	UserCannotUseBurstReactions,
 
 	ApplicationNotYetAvailable = 110_001,
 
@@ -292,6 +293,16 @@ export enum RESTJSONErrorCodes {
 
 	CannotEnableOnboardingRequirementsAreNotMet = 350_000,
 	CannotUpdateOnboardingWhileBelowRequirements,
+
+	FailedToBanUsers = 500_000,
+
+	PollVotingBlocked = 520_000,
+	PollExpired,
+	InvalidChannelTypeForPollCreation,
+	CannotEditAPollMessage,
+	CannotUseAnEmojiIncludedWithThePoll,
+
+	CannotExpireANonPollMessage = 520_006,
 }
 
 /**
@@ -324,6 +335,7 @@ export enum Locale {
 	Romanian = 'ro',
 	Russian = 'ru',
 	SpanishES = 'es-ES',
+	SpanishLATAM = 'es-419',
 	Swedish = 'sv-SE',
 	Thai = 'th',
 	Turkish = 'tr',