+++ rest/v9/oauth2.ts
@@ -122,8 +122,8 @@ export interface RESTOAuth2BotAuthorizationQuery {
 	 */
 	scope:
 		| OAuth2Scopes.Bot
-		| `${string}${' ' | '%20'}${OAuth2Scopes.Bot}`
 		| `${OAuth2Scopes.Bot}${' ' | '%20'}${string}`
+		| `${string}${' ' | '%20'}${OAuth2Scopes.Bot}`
 		| `${string}${' ' | '%20'}${OAuth2Scopes.Bot}${string}${' ' | '%20'}`;
 	/**
 	 * The permissions you're requesting
@@ -151,8 +151,8 @@ export interface RESTOAuth2AdvancedBotAuthorizationQuery {
 	 */
 	scope:
 		| OAuth2Scopes.Bot
-		| `${string}${' ' | '%20'}${OAuth2Scopes.Bot}`
 		| `${OAuth2Scopes.Bot}${' ' | '%20'}${string}`
+		| `${string}${' ' | '%20'}${OAuth2Scopes.Bot}`
 		| `${string}${' ' | '%20'}${OAuth2Scopes.Bot}${string}${' ' | '%20'}`;
 	/**
 	 * The required permissions bitfield, stringified
@@ -196,4 +196,5 @@ export interface RESTPostOAuth2AccessTokenWithBotAndWebhookIncomingScopeResult {
 }
 
 export type RESTPostOAuth2AccessTokenWithBotAndGuildsAndWebhookIncomingScopeResult =
-	RESTPostOAuth2AccessTokenWithBotAndGuildsScopeResult & RESTPostOAuth2AccessTokenWithBotAndWebhookIncomingScopeResult;
+	RESTPostOAuth2AccessTokenWithBotAndGuildsScopeResult &
+		RESTPostOAuth2AccessTokenWithBotAndWebhookIncomingScopeResult;